Decrease code duplication
