+++ datafusion/core/src/physical_plan/repartition/mod.rs
@@ -16,7 +16,8 @@
 // under the License.
 
 //! The repartition operator maps N input partitions to M output partitions based on a
-//! partitioning scheme.
+//! partitioning scheme (according to flag `preserve_order` ordering can be preserved during
+//! repartitioning if its input is ordered).
 
 use std::pin::Pin;
 use std::sync::Arc;
@@ -24,7 +25,9 @@ use std::task::{Context, Poll};
 use std::{any::Any, vec};
 
 use crate::physical_plan::hash_utils::create_hashes;
-use crate::physical_plan::repartition::distributor_channels::channels;
+use crate::physical_plan::repartition::distributor_channels::{
+    channels, partition_aware_channels,
+};
 use crate::physical_plan::{
     DisplayFormatType, EquivalenceProperties, ExecutionPlan, Partitioning, Statistics,
 };
@@ -42,6 +45,9 @@ use super::expressions::PhysicalSortExpr;
 use super::metrics::{self, ExecutionPlanMetricsSet, MetricBuilder, MetricsSet};
 use super::{RecordBatchStream, SendableRecordBatchStream};
 
+use crate::physical_plan::common::transpose;
+use crate::physical_plan::metrics::BaselineMetrics;
+use crate::physical_plan::sorts::streaming_merge;
 use datafusion_execution::TaskContext;
 use datafusion_physical_expr::PhysicalExpr;
 use futures::stream::Stream;
@@ -50,9 +56,10 @@ use hashbrown::HashMap;
 use parking_lot::Mutex;
 
 pub(crate) mod distributor_channels;
-pub mod sort_preserving_repartition;
 
 type MaybeBatch = Option<Result<RecordBatch>>;
+type InputPartitionsToCurrentPartitionSender = Vec<DistributionSender<MaybeBatch>>;
+type InputPartitionsToCurrentPartitionReceiver = Vec<DistributionReceiver<MaybeBatch>>;
 
 /// Inner state of [`RepartitionExec`].
 #[derive(Debug)]
@@ -62,8 +69,8 @@ struct RepartitionExecState {
     channels: HashMap<
         usize,
         (
-            DistributionSender<MaybeBatch>,
-            DistributionReceiver<MaybeBatch>,
+            InputPartitionsToCurrentPartitionSender,
+            InputPartitionsToCurrentPartitionReceiver,
             SharedMemoryReservation,
         ),
     >,
@@ -245,6 +252,9 @@ pub struct RepartitionExec {
 
     /// Execution metrics
     metrics: ExecutionPlanMetricsSet,
+
+    /// Boolean flag to decide whether to preserve ordering
+    preserve_order: bool,
 }
 
 #[derive(Debug, Clone)]
@@ -298,6 +308,15 @@ impl RepartitionExec {
     pub fn partitioning(&self) -> &Partitioning {
         &self.partitioning
     }
+
+    /// Get name of the Executor
+    pub fn name(&self) -> &str {
+        if self.preserve_order {
+            "SortPreservingRepartitionExec"
+        } else {
+            "RepartitionExec"
+        }
+    }
 }
 
 impl ExecutionPlan for RepartitionExec {
@@ -345,8 +364,12 @@ impl ExecutionPlan for RepartitionExec {
     }
 
     fn maintains_input_order(&self) -> Vec<bool> {
-        // We preserve ordering when input partitioning is 1
-        vec![self.input().output_partitioning().partition_count() <= 1]
+        if self.preserve_order {
+            vec![true]
+        } else {
+            // We preserve ordering when input partitioning is 1
+            vec![self.input().output_partitioning().partition_count() <= 1]
+        }
     }
 
     fn equivalence_properties(&self) -> EquivalenceProperties {
@@ -359,7 +382,8 @@ impl ExecutionPlan for RepartitionExec {
         context: Arc<TaskContext>,
     ) -> Result<SendableRecordBatchStream> {
         trace!(
-            "Start RepartitionExec::execute for partition: {}",
+            "Start {}::execute for partition: {}",
+            self.name(),
             partition
         );
         // lock mutexes
@@ -370,13 +394,29 @@ impl ExecutionPlan for RepartitionExec {
 
         // if this is the first partition to be invoked then we need to set up initial state
         if state.channels.is_empty() {
-            // create one channel per *output* partition
-            // note we use a custom channel that ensures there is always data for each receiver
-            // but limits the amount of buffering if required.
-            let (txs, rxs) = channels(num_output_partitions);
+            let (txs, rxs) = if self.preserve_order {
+                let (txs, rxs) =
+                    partition_aware_channels(num_input_partitions, num_output_partitions);
+                // Take transpose of senders and receivers. `state.channels` keeps track of entries per output partition
+                let txs = transpose(txs);
+                let rxs = transpose(rxs);
+                (txs, rxs)
+            } else {
+                // create one channel per *output* partition
+                // note we use a custom channel that ensures there is always data for each receiver
+                // but limits the amount of buffering if required.
+                let (txs, rxs) = channels(num_output_partitions);
+                // Clone sender for ech input partitions
+                let txs = txs
+                    .into_iter()
+                    .map(|item| vec![item; num_input_partitions])
+                    .collect::<Vec<_>>();
+                let rxs = rxs.into_iter().map(|item| vec![item]).collect::<Vec<_>>();
+                (txs, rxs)
+            };
             for (partition, (tx, rx)) in txs.into_iter().zip(rxs).enumerate() {
                 let reservation = Arc::new(Mutex::new(
-                    MemoryConsumer::new(format!("RepartitionExec[{partition}]"))
+                    MemoryConsumer::new(format!("{}[{partition}]", self.name()))
                         .register(context.memory_pool()),
                 ));
                 state.channels.insert(partition, (tx, rx, reservation));
@@ -389,13 +429,13 @@ impl ExecutionPlan for RepartitionExec {
                     .channels
                     .iter()
                     .map(|(partition, (tx, _rx, reservation))| {
-                        (*partition, (tx.clone(), Arc::clone(reservation)))
+                        (*partition, (tx[i].clone(), Arc::clone(reservation)))
                     })
                     .collect();
 
                 let r_metrics = RepartitionMetrics::new(i, partition, &self.metrics);
 
-                let input_task = tokio::spawn(pull_from_input(
+                let input_task = tokio::spawn(Self::pull_from_input(
                     self.input.clone(),
                     i,
                     txs.clone(),
@@ -406,7 +446,7 @@ impl ExecutionPlan for RepartitionExec {
 
                 // In a separate task, wait for each input to be done
                 // (and pass along any errors, including panic!s)
-                let join_handle = tokio::spawn(wait_for_task(
+                let join_handle = tokio::spawn(Self::wait_for_task(
                     AbortOnDropSingle::new(input_task),
                     txs.into_iter()
                         .map(|(partition, (tx, _reservation))| (partition, tx))
@@ -419,24 +459,53 @@ impl ExecutionPlan for RepartitionExec {
         }
 
         trace!(
-            "Before returning stream in RepartitionExec::execute for partition: {}",
+            "Before returning stream in {}::execute for partition: {}",
+            self.name(),
             partition
         );
 
         // now return stream for the specified *output* partition which will
         // read from the channel
-        let (_tx, rx, reservation) = state
+        let (_tx, mut rx, reservation) = state
             .channels
             .remove(&partition)
             .expect("partition not used yet");
-        Ok(Box::pin(RepartitionStream {
-            num_input_partitions,
-            num_input_partitions_processed: 0,
-            schema: self.input.schema(),
-            input: rx,
-            drop_helper: Arc::clone(&state.abort_helper),
-            reservation,
-        }))
+
+        if self.preserve_order {
+            // Store streams from all the input partitions:
+            let input_streams = rx
+                .into_iter()
+                .map(|receiver| {
+                    Box::pin(PerPartitionStream {
+                        schema: self.schema(),
+                        receiver,
+                        drop_helper: Arc::clone(&state.abort_helper),
+                        reservation: reservation.clone(),
+                    }) as SendableRecordBatchStream
+                })
+                .collect::<Vec<_>>();
+            // Note that receiver size (`rx.len()`) and `num_input_partitions` are same.
+
+            // Get existing ordering:
+            let sort_exprs = self.input.output_ordering().unwrap_or(&[]);
+            // Merge streams (while preserving ordering) coming from input partitions to this partition:
+            streaming_merge(
+                input_streams,
+                self.schema(),
+                sort_exprs,
+                BaselineMetrics::new(&self.metrics, partition),
+                context.session_config().batch_size(),
+            )
+        } else {
+            Ok(Box::pin(RepartitionStream {
+                num_input_partitions,
+                num_input_partitions_processed: 0,
+                schema: self.input.schema(),
+                input: rx.swap_remove(0),
+                drop_helper: Arc::clone(&state.abort_helper),
+                reservation,
+            }))
+        }
     }
 
     fn metrics(&self) -> Option<MetricsSet> {
@@ -452,7 +521,8 @@ impl ExecutionPlan for RepartitionExec {
             DisplayFormatType::Default | DisplayFormatType::Verbose => {
                 write!(
                     f,
-                    "RepartitionExec: partitioning={:?}, input_partitions={}",
+                    "{}: partitioning={:?}, input_partitions={}",
+                    self.name(),
                     self.partitioning,
                     self.input.output_partitioning().partition_count()
                 )
@@ -479,7 +549,143 @@ impl RepartitionExec {
                 abort_helper: Arc::new(AbortOnDropMany::<()>(vec![])),
             })),
             metrics: ExecutionPlanMetricsSet::new(),
+            preserve_order: false,
         })
+    }
+
+    /// Set sort information
+    pub fn with_preserve_order(mut self) -> Self {
+        self.preserve_order = true;
+        self
+    }
+
+    /// Pulls data from the specified input plan, feeding it to the
+    /// output partitions based on the desired partitioning
+    ///
+    /// i is the input partition index
+    ///
+    /// txs hold the output sending channels for each output partition
+    async fn pull_from_input(
+        input: Arc<dyn ExecutionPlan>,
+        i: usize,
+        mut txs: HashMap<
+            usize,
+            (DistributionSender<MaybeBatch>, SharedMemoryReservation),
+        >,
+        partitioning: Partitioning,
+        r_metrics: RepartitionMetrics,
+        context: Arc<TaskContext>,
+    ) -> Result<()> {
+        let mut partitioner =
+            BatchPartitioner::try_new(partitioning, r_metrics.repart_time.clone())?;
+
+        // execute the child operator
+        let timer = r_metrics.fetch_time.timer();
+        let mut stream = input.execute(i, context)?;
+        timer.done();
+
+        // While there are still outputs to send to, keep
+        // pulling inputs
+        let mut batches_until_yield = partitioner.num_partitions();
+        while !txs.is_empty() {
+            // fetch the next batch
+            let timer = r_metrics.fetch_time.timer();
+            let result = stream.next().await;
+            timer.done();
+
+            // Input is done
+            let batch = match result {
+                Some(result) => result?,
+                None => break,
+            };
+
+            for res in partitioner.partition_iter(batch)? {
+                let (partition, batch) = res?;
+                let size = batch.get_array_memory_size();
+
+                let timer = r_metrics.send_time.timer();
+                // if there is still a receiver, send to it
+                if let Some((tx, reservation)) = txs.get_mut(&partition) {
+                    reservation.lock().try_grow(size)?;
+
+                    if tx.send(Some(Ok(batch))).await.is_err() {
+                        // If the other end has hung up, it was an early shutdown (e.g. LIMIT)
+                        reservation.lock().shrink(size);
+                        txs.remove(&partition);
+                    }
+                }
+                timer.done();
+            }
+
+            // If the input stream is endless, we may spin forever and
+            // never yield back to tokio.  See
+            // https://github.com/apache/arrow-datafusion/issues/5278.
+            //
+            // However, yielding on every batch causes a bottleneck
+            // when running with multiple cores. See
+            // https://github.com/apache/arrow-datafusion/issues/6290
+            //
+            // Thus, heuristically yield after producing num_partition
+            // batches
+            //
+            // In round robin this is ideal as each input will get a
+            // new batch. In hash partitioning it may yield too often
+            // on uneven distributions even if some partition can not
+            // make progress, but parallelism is going to be limited
+            // in that case anyways
+            if batches_until_yield == 0 {
+                tokio::task::yield_now().await;
+                batches_until_yield = partitioner.num_partitions();
+            } else {
+                batches_until_yield -= 1;
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Waits for `input_task` which is consuming one of the inputs to
+    /// complete. Upon each successful completion, sends a `None` to
+    /// each of the output tx channels to signal one of the inputs is
+    /// complete. Upon error, propagates the errors to all output tx
+    /// channels.
+    async fn wait_for_task(
+        input_task: AbortOnDropSingle<Result<()>>,
+        txs: HashMap<usize, DistributionSender<MaybeBatch>>,
+    ) {
+        // wait for completion, and propagate error
+        // note we ignore errors on send (.ok) as that means the receiver has already shutdown.
+        match input_task.await {
+            // Error in joining task
+            Err(e) => {
+                let e = Arc::new(e);
+
+                for (_, tx) in txs {
+                    let err = Err(DataFusionError::Context(
+                        "Join Error".to_string(),
+                        Box::new(DataFusionError::External(Box::new(Arc::clone(&e)))),
+                    ));
+                    tx.send(Some(err)).await.ok();
+                }
+            }
+            // Error from running input task
+            Ok(Err(e)) => {
+                let e = Arc::new(e);
+
+                for (_, tx) in txs {
+                    // wrap it because need to send error to all output partitions
+                    let err = Err(DataFusionError::External(Box::new(e.clone())));
+                    tx.send(Some(err)).await.ok();
+                }
+            }
+            // Input task completed successfully
+            Ok(Ok(())) => {
+                // notify each output partition that this input partition has no more data
+                for (_, tx) in txs {
+                    tx.send(None).await.ok();
+                }
+            }
+        }
     }
 }
 
@@ -551,129 +757,53 @@ impl RecordBatchStream for RepartitionStream {
     }
 }
 
-/// Pulls data from the specified input plan, feeding it to the
-/// output partitions based on the desired partitioning
-///
-/// i is the input partition index
-///
-/// txs hold the output sending channels for each output partition
-async fn pull_from_input(
-    input: Arc<dyn ExecutionPlan>,
-    i: usize,
-    mut txs: HashMap<usize, (DistributionSender<MaybeBatch>, SharedMemoryReservation)>,
-    partitioning: Partitioning,
-    r_metrics: RepartitionMetrics,
-    context: Arc<TaskContext>,
-) -> Result<()> {
-    let mut partitioner =
-        BatchPartitioner::try_new(partitioning, r_metrics.repart_time.clone())?;
+/// This struct converts a receiver to a stream.
+/// Receiver receives data on an SPSC channel.
+struct PerPartitionStream {
+    /// Schema wrapped by Arc
+    schema: SchemaRef,
 
-    // execute the child operator
-    let timer = r_metrics.fetch_time.timer();
-    let mut stream = input.execute(i, context)?;
-    timer.done();
+    /// channel containing the repartitioned batches
+    receiver: DistributionReceiver<MaybeBatch>,
 
-    // While there are still outputs to send to, keep
-    // pulling inputs
-    let mut batches_until_yield = partitioner.num_partitions();
-    while !txs.is_empty() {
-        // fetch the next batch
-        let timer = r_metrics.fetch_time.timer();
-        let result = stream.next().await;
-        timer.done();
+    /// Handle to ensure background tasks are killed when no longer needed.
+    #[allow(dead_code)]
+    drop_helper: Arc<AbortOnDropMany<()>>,
 
-        // Input is done
-        let batch = match result {
-            Some(result) => result?,
-            None => break,
-        };
-
-        for res in partitioner.partition_iter(batch)? {
-            let (partition, batch) = res?;
-            let size = batch.get_array_memory_size();
-
-            let timer = r_metrics.send_time.timer();
-            // if there is still a receiver, send to it
-            if let Some((tx, reservation)) = txs.get_mut(&partition) {
-                reservation.lock().try_grow(size)?;
-
-                if tx.send(Some(Ok(batch))).await.is_err() {
-                    // If the other end has hung up, it was an early shutdown (e.g. LIMIT)
-                    reservation.lock().shrink(size);
-                    txs.remove(&partition);
-                }
-            }
-            timer.done();
-        }
-
-        // If the input stream is endless, we may spin forever and
-        // never yield back to tokio.  See
-        // https://github.com/apache/arrow-datafusion/issues/5278.
-        //
-        // However, yielding on every batch causes a bottleneck
-        // when running with multiple cores. See
-        // https://github.com/apache/arrow-datafusion/issues/6290
-        //
-        // Thus, heuristically yield after producing num_partition
-        // batches
-        //
-        // In round robin this is ideal as each input will get a
-        // new batch. In hash partitioning it may yield too often
-        // on uneven distributions even if some partition can not
-        // make progress, but parallelism is going to be limited
-        // in that case anyways
-        if batches_until_yield == 0 {
-            tokio::task::yield_now().await;
-            batches_until_yield = partitioner.num_partitions();
-        } else {
-            batches_until_yield -= 1;
-        }
-    }
-
-    Ok(())
+    /// Memory reservation.
+    reservation: SharedMemoryReservation,
 }
 
-/// Waits for `input_task` which is consuming one of the inputs to
-/// complete. Upon each successful completion, sends a `None` to
-/// each of the output tx channels to signal one of the inputs is
-/// complete. Upon error, propagates the errors to all output tx
-/// channels.
-async fn wait_for_task(
-    input_task: AbortOnDropSingle<Result<()>>,
-    txs: HashMap<usize, DistributionSender<MaybeBatch>>,
-) {
-    // wait for completion, and propagate error
-    // note we ignore errors on send (.ok) as that means the receiver has already shutdown.
-    match input_task.await {
-        // Error in joining task
-        Err(e) => {
-            let e = Arc::new(e);
+impl Stream for PerPartitionStream {
+    type Item = Result<RecordBatch>;
 
-            for (_, tx) in txs {
-                let err = Err(DataFusionError::Context(
-                    "Join Error".to_string(),
-                    Box::new(DataFusionError::External(Box::new(Arc::clone(&e)))),
-                ));
-                tx.send(Some(err)).await.ok();
+    fn poll_next(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Option<Self::Item>> {
+        match self.receiver.recv().poll_unpin(cx) {
+            Poll::Ready(Some(Some(v))) => {
+                if let Ok(batch) = &v {
+                    self.reservation
+                        .lock()
+                        .shrink(batch.get_array_memory_size());
+                }
+                Poll::Ready(Some(v))
             }
+            Poll::Ready(Some(None)) => {
+                // Input partition has finished sending batches
+                Poll::Ready(None)
+            }
+            Poll::Ready(None) => Poll::Ready(None),
+            Poll::Pending => Poll::Pending,
         }
-        // Error from running input task
-        Ok(Err(e)) => {
-            let e = Arc::new(e);
+    }
+}
 
-            for (_, tx) in txs {
-                // wrap it because need to send error to all output partitions
-                let err = Err(DataFusionError::External(Box::new(e.clone())));
-                tx.send(Some(err)).await.ok();
-            }
-        }
-        // Input task completed successfully
-        Ok(Ok(())) => {
-            // notify each output partition that this input partition has no more data
-            for (_, tx) in txs {
-                tx.send(None).await.ok();
-            }
-        }
+impl RecordBatchStream for PerPartitionStream {
+    /// Get the schema
+    fn schema(&self) -> SchemaRef {
+        self.schema.clone()
     }
 }
